🥒 feat: change to get FOLDER_ID from const to environment variables

+++ packages/hello/src/index.ts
@@ -17,23 +17,27 @@ function getSpreadsheetName(hoge: string) {
   return `${SpreadsheetApp?.getActiveSpreadsheet()?.getName()} ${hoge}`;
 }
 
-// IDをどのように知るかと言いますと、フォルダであればGoogleドライブでそのフォルダを開いている際のURLの
-// https://drive.google.com/drive/folders/{フォルダID}
-// の部分です。
-// Google Apps ScriptでGoogleドライブを操作する最も簡単なスクリプト - https://tonari-it.com/gas-google-drive-app/
-const GAS_FOLDER_ID = '1S4-Wz1p5GLyDm9i42lFZ3iSj2ofQOitU';
 function creatSS() {
   // スプレッドシートオブジェクトを作成する
   const id = SpreadsheetApp.create('new Spread Sheet').getId();
   // スプレッドシートのファイルオブジェクトを取得する
   const file = DriveApp.getFileById(id);
 
-  // スプレッドシートの移動先となるフォルダー名
-  const folder = DriveApp.getFolderById(GAS_FOLDER_ID);
+  // IDをどのように知るかと言いますと、フォルダであればGoogleドライブでそのフォルダを開いている際のURLの
+  // https://drive.google.com/drive/folders/{フォルダID}
+  // の部分です。
+  // Google Apps ScriptでGoogleドライブを操作する最も簡単なスクリプト - https://tonari-it.com/gas-google-drive-app/
+  // スプレッドシートの移動先となるフォルダー名（環境変数より取得）
+  const GAS_FOLDER_ID: string | null = PropertiesService.getScriptProperties().getProperty('GAS_FOLDER_ID');
 
-  // ルートに出来たスプレッドシートをフォルダーに移動して、ルートにある同ファイルを削除する
-  folder.addFile(file);
-  DriveApp.getRootFolder().removeFile(file);
+  let folder;
+  if (GAS_FOLDER_ID) {
+    folder = DriveApp.getFolderById(GAS_FOLDER_ID);
 
-  console.log(folder.getName());
+    // ルートに出来たスプレッドシートをフォルダーに移動して、ルートにある同ファイルを削除する
+    folder.addFile(file);
+    DriveApp.getRootFolder().removeFile(file);
+
+    console.log(folder.getName());
+  }
 }
